+++ CHANGES.rst
@@ -6,6 +6,7 @@ ChangeLog
 ------------------
 
 * Validate CLI user inputs and provides hints.
+* Abandon branches of the search space as soon as possible. Closes #3.
 
 
 0.8.0 (2015-08-15)